# docs/swagger.yaml
definitions:
...
  controllers.orderDTO:
    properties:
      customer_id:
        type: integer
      price:
        type: number
      product_id:
        type: integer
      quantity:
        type: integer
      total_amount:
        type: number
    type: object
  controllers.paymentDTO:
    properties:
      amount:
        type: number
      uid_order:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /order:
    get:
      description: Get a list of Order.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get all Order.
      tags:
      - Order
  /order_item:
    get:
      description: Get a list of OrderItem.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get all OrderItems.
      tags:
      - Order
  /orders:
    post:
      description: Creating a new Order.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create a new Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.orderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Create Order.
      tags:
      - Order
  /orders/{id}:
    delete:
      description: Delete a Order by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Order id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Delete one Order.
      tags:
      - Order
    put:
      description: Update Order by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: order id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an Order
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.orderDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Update Order.
      tags:
      - Order
  /payments:
    get:
      description: Get a list of payments.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get all payments.
      tags:
      - Payments
    post:
      description: Creating a new payment.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: the body to create a new payments
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.paymentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Create New payment.
      tags:
      - Payments
  /payments/{id}:
    delete:
      description: Delete a payment by id.
      parameters:
      - description: payment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Delete one payment.
      tags:
      - Payments
    get:
      description: Get a payments by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: payments id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get payment by id
      tags:
      - Payments
    put:
      description: Update payments by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: payments id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update an payments
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.paymentDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Update payments.
      tags:
      - Payments
  /payments/{id}/orders:
    get:
      description: Get all Orders by payments Id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: payment id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerToken: []
      summary: Get Orders By Payment ID.
      tags:
      - Payments
...
swagger: "2.0"